+++ src/components/feature/LeftoverList.tsx
@@ -2,12 +2,12 @@ import { clsx } from "clsx"
 import dayjs from "dayjs"
 import { useCallback, useMemo } from "react"
 
+import { dayjsToKey, keyToDayjs, Task } from "../../model/task"
+import { useRefreshTaskListByDate } from "../../model/task-list-by-date"
 import {
   useLeftoverTaskList,
   useRefreshLeftoverTaskList,
-} from "../../model/leftover-task-list-query"
-import { dayjsToKey, keyToDayjs, Task } from "../../model/task"
-import { useRefreshTaskListByDate } from "../../model/task-list-by-date"
+} from "../../model/task-list-leftovers"
 import { useRefreshTaskQuery } from "../../model/task-query"
 import { moveTaskToday, removeTask } from "../../model/task-usecase"
 import { Button } from "../ui/Button"
