+++ src/components/pages/Yarukoto/Yarukoto.tsx
@@ -4,12 +4,12 @@ import React, { useCallback, useState } from "react"
 import { Link } from "rocon/react"
 
 import { routes_y } from "../../../app/Router"
-import { useRefreshLeftoverTaskList } from "../../../model/leftover-task-list-query"
 import { DateKey, dayjsToKey, keyToDayjs } from "../../../model/task"
 import {
   useRefreshTaskListByDate,
   useTaskListByDate,
 } from "../../../model/task-list-by-date"
+import { useRefreshLeftoverTaskList } from "../../../model/task-list-leftovers"
 import { useRefreshTaskQuery } from "../../../model/task-query"
 import { useSortableList } from "../../../model/task-sortable-list"
 import {
